squash: 4 contiguous commits by Tura AI
Contiguous-author squash preserving the final tree of the original chain.
Original range: 2e685722faf4b55e1f6e804bb83159a5b963f2a5..aa0004844c2eb883d122ed1419ea3eabbd4d8ec3

Original commits:
- 2e685722faf4b55e1f6e804bb83159a5b963f2a5 tura completed tura-1784847285472-c901ef10-6ded-4fe3-a1ff-936341095eae: Git branch diagnostics
- 65a5a82ead79e7f1e1cf3bc3f061a193bd358cb3 tura completed tura-1784847285472-c901ef10-6ded-4fe3-a1ff-936341095eae: Runtime staging diagnostics
- a430211031677d794a7faf7b4639dd5e2a61c048 tura completed tura-1784848614767-85776279-fc57-4245-a93c-109449726884: TUI live rendering
- aa0004844c2eb883d122ed1419ea3eabbd4d8ec3 tura completed tura-1784848614767-85776279-fc57-4245-a93c-109449726884: TUI live rendering

+++ crates/runtime/tests/business/session_log_queue_recovery_flow.rs
@@ -120,13 +120,13 @@ fn runtime_session_log_business_flow_drains_concurrent_offline_writes_without_lo
                 &session_id,
                 &workspace_text,
                 "Runtime Queue Recovery",
-                index as i64,
+                index,
             ))?;
             typed_session::enqueue_delta(
                 &session_id,
                 &workspace_text,
                 "Runtime Queue Recovery",
-                index as i64,
+                index,
                 typed_session::entries_from_messages(
                     0,
                     vec![message_payload(
